backend/requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==2.2.19
django-colorfield==0.7.2
django-cors-headers==3.0.0
django-debug-toolbar==3.2.4
django-extra-fields==3.0.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Faker==12.0.1
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==1.7.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.1.2
mypy==0.971
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.3.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
reportlab==3.6.12
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
